requirements.txt
scapy
matplotlib
numpy
scipy
hdrhistogram
git+git://github.com/theodorsm/hdr-plot.git@master
python-dotenv
